+++ pkg/morph/event/listener.go
@@ -6,21 +6,16 @@ import (
 	"fmt"
 	"sync"
 
+	"github.com/nspcc-dev/neo-go/pkg/core/block"
 	"github.com/nspcc-dev/neo-go/pkg/core/state"
 	"github.com/nspcc-dev/neo-go/pkg/neorpc/result"
 	"github.com/nspcc-dev/neo-go/pkg/util"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
-	"github.com/panjf2000/ants/v2"
 	"go.uber.org/zap"
 )
 
 // Listener is an interface of smart contract notification event listener.
 type Listener interface {
-	// Listen must start the event listener.
-	//
-	// Must listen to events with the parser installed.
-	Listen(context.Context)
-
 	// ListenWithError must start the event listener.
 	//
 	// Must listen to events with the parser installed.
@@ -69,12 +64,12 @@ type Listener interface {
 	// Has no effect if EnableNotarySupport was not called before Listen or ListenWithError.
 	RegisterNotaryHandler(NotaryHandlerInfo)
 
-	// RegisterBlockHandler must register chain block handler.
+	// RegisterHeaderHandler must register chain header handler.
 	//
-	// The specified handler must be called after each capture and parsing of the new block from chain.
+	// The specified handler will be called after each capture and parsing of the new header from chain.
 	//
 	// Must ignore nil handlers.
-	RegisterBlockHandler(BlockHandler)
+	RegisterHeaderHandler(HeaderHandler)
 
 	// Stop must stop the event listener.
 	Stop()
@@ -86,8 +81,6 @@ type ListenerParams struct {
 	Logger *zap.Logger
 
 	Client *client.Client
-
-	WorkerPoolCapacity int
 }
 
 type listener struct {
@@ -110,33 +103,19 @@ type listener struct {
 
 	cli *client.Client
 
-	blockHandlers []BlockHandler
-
-	pool *ants.Pool
+	headerHandlers []HeaderHandler
 }
 
 const newListenerFailMsg = "could not instantiate Listener"
+
+// notaryRoutineNum limits the number of concurrent notary request handlers.
+const notaryRoutineNum = 10
 
 var (
 	errNilLogger = errors.New("nil logger")
 
 	errNilSubscriber = errors.New("nil event client")
 )
-
-// Listen starts the listening for events with registered handlers.
-//
-// Executes once, all subsequent calls do nothing.
-//
-// Returns an error if listener was already started.
-func (l *listener) Listen(ctx context.Context) {
-	l.startOnce.Do(func() {
-		if err := l.listen(ctx); err != nil {
-			l.log.Error("could not start listen to events",
-				zap.Error(err),
-			)
-		}
-	})
-}
 
 // ListenWithError starts the listening for events with registered handlers and
 // passing error message to intError channel if subscriber channel has been closed.
@@ -147,9 +126,6 @@ func (l *listener) Listen(ctx context.Context) {
 func (l *listener) ListenWithError(ctx context.Context, intError chan<- error) {
 	l.startOnce.Do(func() {
 		if err := l.listen(ctx); err != nil {
-			l.log.Error("could not start listen to events",
-				zap.Error(err),
-			)
 			intError <- err
 		}
 	})
@@ -159,14 +135,24 @@ func (l *listener) listen(ctx context.Context) error {
 	// mark listener as started
 	l.started = true
 
-	subErrCh := make(chan error)
+	ctx, cancel := context.WithCancelCause(ctx)
+	notifyCh, headerCh, notaryCh := l.cli.Notifications()
 
-	go l.subscribe(subErrCh)
-
-	return l.listenLoop(ctx, subErrCh)
+	go l.subscribe(cancel)
+	go l.listenForHeaders(ctx, cancel, headerCh)
+	go l.listenForNotary(ctx, cancel, notaryCh)
+	// We need to block this goroutine too
+	l.listenForNotifications(ctx, cancel, notifyCh)
+	err := context.Cause(ctx)
+	if !errors.Is(err, context.Canceled) {
+		// Some real cause.
+		return err
+	}
+	// Upper-layer cancellation.
+	return nil
 }
 
-func (l *listener) subscribe(errCh chan error) {
+func (l *listener) subscribe(cancel context.CancelCauseFunc) {
 	// create the list of listening contract hashes
 	hashes := make([]util.Uint160, 0)
 
@@ -188,84 +174,94 @@ func (l *listener) subscribe(errCh chan error) {
 
 	err := l.cli.ReceiveExecutionNotifications(hashes)
 	if err != nil {
-		errCh <- fmt.Errorf("could not subscribe for notifications: %w", err)
+		cancel(err)
 		return
 	}
 
-	if len(l.blockHandlers) > 0 {
-		if err = l.cli.ReceiveBlocks(); err != nil {
-			errCh <- fmt.Errorf("could not subscribe for blocks: %w", err)
+	if len(l.headerHandlers) > 0 {
+		if err = l.cli.ReceiveHeaders(); err != nil {
+			cancel(err)
 			return
 		}
 	}
 
 	if l.listenNotary {
 		if err = l.cli.ReceiveNotaryRequests(l.notaryMainTXSigner); err != nil {
-			errCh <- fmt.Errorf("could not subscribe for notary requests: %w", err)
+			cancel(err)
 			return
 		}
 	}
 }
 
-func (l *listener) listenLoop(ctx context.Context, subErrCh chan error) error {
-	nCh, bCh, notaryCh := l.cli.Notifications()
-	var res error
+func (l *listener) listenForHeaders(ctx context.Context, cancel context.CancelCauseFunc, headerCh <-chan *block.Header) {
+	for {
+		select {
+		case <-ctx.Done():
+			l.log.Info("header listener stopped",
+				zap.Error(context.Cause(ctx)))
+			return
+		case h, ok := <-headerCh:
+			if !ok {
+				l.log.Warn("header channel closed, stopping event listener")
+				cancel(errors.New("header notification channel was closed"))
+				return
+			}
+
+			for i := range l.headerHandlers {
+				l.headerHandlers[i](h)
+			}
+		}
+	}
+}
+
+func (l *listener) listenForNotary(ctx context.Context, cancel context.CancelCauseFunc, notaryCh <-chan *result.NotaryRequestEvent) {
+	var secondaryCh = make(chan *result.NotaryRequestEvent)
+
+	for range notaryRoutineNum {
+		go func() {
+			for notaryEvent := range secondaryCh {
+				l.parseAndHandleNotary(notaryEvent)
+			}
+		}()
+	}
 
 loop:
 	for {
 		select {
-		case res = <-subErrCh:
-			l.log.Error("stop event listener by error", zap.Error(res))
-			break loop
 		case <-ctx.Done():
-			l.log.Info("stop event listener by context",
-				zap.String("reason", ctx.Err().Error()),
-			)
+			l.log.Info("notary listener stopped",
+				zap.Error(context.Cause(ctx)))
 			break loop
-		case notifyEvent, ok := <-nCh:
-			if !ok {
-				l.log.Warn("stop event listener by notification channel")
-				res = errors.New("event subscriber connection has been terminated")
-				break loop
-			}
-
-			if err := l.pool.Submit(func() {
-				l.parseAndHandleNotification(notifyEvent)
-			}); err != nil {
-				l.log.Warn("listener worker pool drained",
-					zap.Int("capacity", l.pool.Cap()))
-			}
 		case notaryEvent, ok := <-notaryCh:
 			if !ok {
-				l.log.Warn("stop event listener by notary channel")
-				res = errors.New("notary event subscriber connection has been terminated")
+				l.log.Warn("notary channel closed, stopping event listener")
+				cancel(errors.New("notary notification channel was closed"))
 				break loop
 			}
 
-			if err := l.pool.Submit(func() {
-				l.parseAndHandleNotary(notaryEvent)
-			}); err != nil {
-				l.log.Warn("listener worker pool drained",
-					zap.Int("capacity", l.pool.Cap()))
-			}
-		case b, ok := <-bCh:
-			if !ok {
-				l.log.Warn("stop event listener by block channel")
-				res = errors.New("new block notification channel is closed")
-				break loop
-			}
-
-			if err := l.pool.Submit(func() {
-				for i := range l.blockHandlers {
-					l.blockHandlers[i](b)
-				}
-			}); err != nil {
-				l.log.Warn("listener worker pool drained",
-					zap.Int("capacity", l.pool.Cap()))
-			}
+			secondaryCh <- notaryEvent
 		}
 	}
-	return res
+	close(secondaryCh)
+}
+
+func (l *listener) listenForNotifications(ctx context.Context, cancel context.CancelCauseFunc, notifyCh <-chan *state.ContainedNotificationEvent) {
+	for {
+		select {
+		case <-ctx.Done():
+			l.log.Info("notification listener stopped",
+				zap.Error(context.Cause(ctx)))
+			return
+		case notifyEvent, ok := <-notifyCh:
+			if !ok {
+				l.log.Warn("notification channel closed, stopping event listener")
+				cancel(errors.New("notification channel was closed"))
+				return
+			}
+
+			l.parseAndHandleNotification(notifyEvent)
+		}
+	}
 }
 
 func (l *listener) parseAndHandleNotification(notifyEvent *state.ContainedNotificationEvent) {
@@ -558,22 +554,17 @@ func (l *listener) Stop() {
 	})
 }
 
-func (l *listener) RegisterBlockHandler(handler BlockHandler) {
+func (l *listener) RegisterHeaderHandler(handler HeaderHandler) {
 	if handler == nil {
-		l.log.Warn("ignore nil block handler")
+		l.log.Warn("ignore nil header handler")
 		return
 	}
 
-	l.blockHandlers = append(l.blockHandlers, handler)
+	l.headerHandlers = append(l.headerHandlers, handler)
 }
 
 // NewListener create the notification event listener instance and returns Listener interface.
 func NewListener(p ListenerParams) (Listener, error) {
-	// defaultPoolCap is a default worker
-	// pool capacity if it was not specified
-	// via params
-	const defaultPoolCap = 100
-
 	switch {
 	case p.Logger == nil:
 		return nil, fmt.Errorf("%s: %w", newListenerFailMsg, errNilLogger)
@@ -581,21 +572,10 @@ func NewListener(p ListenerParams) (Listener, error) {
 		return nil, fmt.Errorf("%s: %w", newListenerFailMsg, errNilSubscriber)
 	}
 
-	poolCap := p.WorkerPoolCapacity
-	if poolCap == 0 {
-		poolCap = defaultPoolCap
-	}
-
-	pool, err := ants.NewPool(poolCap, ants.WithNonblocking(true))
-	if err != nil {
-		return nil, fmt.Errorf("could not init worker pool: %w", err)
-	}
-
 	return &listener{
 		notificationParsers:  make(map[scriptHashWithType]NotificationParser),
 		notificationHandlers: make(map[scriptHashWithType][]Handler),
 		log:                  p.Logger,
 		cli:                  p.Client,
-		pool:                 pool,
 	}, nil
 }